Mark some attestation errors as not retriable (#2130)
* Mark some attestation errors as not retriable

I noticed in testing that fog-distro retries infinitely on report
verification failures, which occur if you are using a slam with bad
sigstruct for the network.

This is partly because slam ignores whether erros are retriable.
But even thick client will retry on all attestation errors, because
the `should_retry` function considers all attestation errors retriable

This makes the `should_retry` logic more nuanced. Attestation errors
now also expose `should_retry`. These errors are now considered
non-retriable, and all others are still retriable.

* Any attestation verification failure (because,
the verifier will be the same next time and the report will be too,
so it has no chance to succeed on retry)
* The responder id construction / uri parsing errors, because, the
uri will not be different next time so it has chance to succeed on
retry.

* fix fog enclave connection error

* fix build

* more fixups

* more fixups