fix mc-crypto-keys build
When setting up another repo, I discovered that this crate doesn't
build properly.

```
error[E0425]: cannot find function `encode_config` in crate `base64`
   --> mobilecoin/crypto/keys/src/x25519.rs:240:31
    |
240 |         let encoded = base64::encode_config(&der, B64_CONFIG);
    |                               ^^^^^^^^^^^^^ not found in `base64`

```

This happens because the `base64::encode_config` function is only
exposed when at least the `alloc` feature is selected. This commit
ensures that it is selected whenever this crate is built.