Create mc-account-keys-types, break dep alloc in mc-crypto-ring-signature
This adds a trait for a subset of the functionality of `PublicAddress` to
a trait called `RingCtAddress`. This is then used in `mc-crypto-ring-signature`
which allows us to break the dependency of that crate on `mc-account-keys`.

This is helpful because `mc-account-keys` depends on `alloc`, but we want
`mc-crypto-ring-signature` not to depend on `alloc`.