Add support for `digestible(transparent)` on rust enums
The intention is that transparent enums are not incorporated into
the hash, and digestible tries to jump directly to the data in the
enum.

The point of this is to allow for schema evolutions where a struct
member is replaced with an enum, one of the options for which is
the previous type. This is a valid way to evolve protobufs, but
it's not currently compatible with digestible, until this commit.

This is being pulled out from the masked-amount-v2 PR to ease review.
We also added a bunch of unit tests and schema evolution tests in
this commit.