remove `InputRuleVerificationData`
This turns out not to be needed, and is actually harmful. The enclave
wants to know that the amount shared secret signed by the originator
of a fractional output matches the amount shared secret used in the
corresponding real output. Therefore it can simply check that directly,
taking the amount shared secrets from the fractional outputs, rather
than making the counterparty copy this to the input rules verification
data.