Simplify the situation around ProposeTxResult vs. TransactionValidationError
Currently, we support converting `TransactionValidationError` to `ProposeTxResult`,
which is necessary to convert errors returned by the enclaves to errors
that can go into a grpc response.

But, we also support the reverse conversion, for no particularly good reason.
`ThickClient` exercises the reverse conversion, so that non-zero `ProposeTxResult`
get converted in a transaction-core / enclave error type when they are returned
to the caller.

This conversion creates a bunch of complexity because it seems bad for it to be
lossy, but making it not lossy is very complicated and imposes constraints on
what kinds of data these errors can contain. It is better if the transaction-core
code can do whatever it wants with it's error types, without regard to concerns
about marshalling them into protobuf, which should be a higher-level concern.

The idea here is:
* Remove the conversion from `TransactionValidationError` back into `ProposeTxResult`.
  Only the one direction is supported.
* The consensus client api service now returns both the `ProposeTxResult` status code
  and an error message, which is the display-string of the original error, in the
  protobuf response. This is a non-breaking change.
* On the thick client side, instead of trying to reconstruct the `TransactionValidationError`,
  we just return the `ProposeTxResult` and the error string.

This promotes a separation of concerns -- now the transaction-core error code
doesn't need to support any particular conversions to support deserializing it.

This also frees us to make the `ProposeTxResult` status code a little bit less granular if
we want. Developers can always get all the details from the err_msg string to debug something,
and we don't have to create more status codes for every possible transaction validation error,
unless we think the client will actually handle those failure modes differently.
It also means that we can put data in the transaction validation errors even if we don't
have a good way to reconstruct it from `ProposeTxResult`.