add MC_CHAIN_ID to the helm templates
This is based on the diff in https://github.com/mobilecoinfoundation/mobilecoin/commit/8203e7cf7e4c275486a3e8676838427d13cd03e0

However, in most places I used the same `valueFrom:` segment as
is used for `MC_BRANCH`, because in that diff those things match also.

I am not sure if that is correct but it seems like it's worth a shot.
Needs review. Thanks.