Update Receipt proto schema for MaskedAmountV2 changes
After masked amount v2, MaskedAmounts are always stored in a proto
oneof, which their versioning information is derived from. It's
generally going to be a bug if a MaskedAmount is not stored in a
oneof this way, because the recipient will not be able to infer
which version it is.

Brian ran into this when updating full-service. It likely will impact
SDKs also.