Add check to mint client to ensure public address doesn't have fog
The mint client reads from a b58 address when it creates a MintTxPrefix,
but it doesn't (before this commit) check if that is actually a fog
address. If it is, the fog part will get silently removed, and then
the fog user won't find it.

This change makes it give an error message.