Add account key abstractions
In order to make progress on the hardware wallets project, we need
to start adding abstractions for more of the functionality of account
keys beyond just signing rings.

One of the main asks is also, the memo signing part. In this commit
we also threw in the key images.

We couldn't throw in Public Address computations yet because it causes
circular dependencies, unless we do more refactoring. We decided to hold
off before this PR gets too large.

Without public addresses, we can't easily compute the short address hash
of the source address, which is needed for the memo computations. So
we can't make the transaction builder part generic over this new memo trait
just yet either.

Nevertheless, committing to this API can help enable further progress on
the hardware wallet side.

One thing I'm anticipating is, if the memo builder contains a reference to
"handle to hardware wallet", and the ring signer is also "handle to hardware
wallet", then we may want to have an `Rc`-wrapper in there to allow it to be
in "two places at once". This shouldn't pose a problem on the
desktop/transaction builder side.

One thing I'm unsure about is if I'm making a mistake making traits like
`RingSigner` take `&self` instead of `&mut self`. We could probably easily
correct course if we have to though. It seems likely that whatever handle
to the hardware wallet we obtain, it may not be perfectly thread-safe, so
maybe `&mut self` is better. OTOH I'd kind of like to avoid complicating
the transaction builder, and using either `Mutex` or `Rc` or something to
resolve the mutability issue.