initial attempt at GHA config that tests android-bindings on PRs (#2543)
Add CI actions that pull the full-service and android-bindings repositories, and then attempt to uprev the mobilecoin submodule to see if changes in a PR have broken it. This should allow us to catch non-backwards-compatible API changes quicker, and hopefully act on them faster so that they don't pile up.