Create recovery_db method that queries expired ingest invocations (#2580)
<!-- List changes here -->

### Motivation

As described in #2337, we'd like to create a tool that cleans up expired egress keys. In effect, this means that we'd like to identify ingest invocations, which are associated with egress keys, that have a `'last_active_at` timestamp before some specified time. 

This PR adds the `RecoveryDb` method and corresponding `SqlRecoveryDb` implementation that allows us to identify expired egress keys.

RFC: This is the first time, to my knowledge, that we are handling diesel / PSQL timestamps and specifying them in the `RecoveryDb` trait, which is `no_std`. I chose to use `chrono::NaiveDateTime` because it's one of the types that `diesel` provides a `ToSql` conversion for AND it doesn't require `std`. Let me know your thoughts.

### Future Work
* Create an executable that uses this method to either print the expired keys or delete them from the DB.