WIP burn-script
This as a similar interface to `drain_accounts.py` and can be used
in a flow where we create wrapped tokens for one of the bootstrapped
sample keys, and then burn those tokens, as part of mint auditor
testing.

This is derived from a script that Eran wrote.

Not tested yet, will likely need fixes.