fixup to partial fills -- input rules check for zero pf change
this was suggested in review of the mcip, and brings the implementation
in line with the spec. Previously this was only checked in the SCI
validate routine.

It doesn't create a bug in the enclave to omit
this check, but it also simplifies review if we just have one spec
for everywhere and enforce it everywhere.