Mitigate tob-5 by making a v2 of client_propose_tx endpoint
TOB-5 is about a "token id oracle" described by TOB during review
of confidential tokens design. The idea is that the consensus enclave
fee map can be changed by the untrusted side, by restarting the node.
If the attacker controls a node and is able to get the user to submit
their Tx once with one fee map, and once with another, then the
priority value leaks information about what the token id of the Tx is.

In principle they could do this without the user knowing -- they would
restart the node and it would not be able to attest to the peers after
changing the fee map, but it could still accept Tx proposals and see
the WellFormedTxContext which conatins the priority number.
Then they turn the node back to normal. The user's app retries when
their transaction doesn't land, and the same Tx comes back to their node,
now with the regular fee map.

This requires having root on a consensus node, but the threat model is
that attackers with root on the consensus node should not be able to
see the user's token id.

We view this as a time-of-check-vs-time-of-use (TOCTOU) bug, and we
try to fix it a way that those bugs are sometimes fixed. The user,
who checked the fee map, also makes an assertion about what the fee-map
is when they try to use the system (by submitting a Tx). Before acting
on the data (which the attacker has potentially tampered with), the
enclave checks if the user's assertion is correct, and rejects the Tx
without creating a WellFormedTxContext if the assertion is not correct.

This should be straightforward to add this check, but unfortunately,
the status quo `client_propose_tx` has a design flaw. It does not
have a `Request` protobuf as the RPC argument, the argument it
takes is directly an (encrypted) Tx. So, we cannot add any additional
data to that protobuf in a non-breaking way without adding things to
`Tx`. And that is untenable in this case because it adds complexity
in sensitive areas of the code. We really just want to add this check
next to the `Tx`.

To fix this, we introduce a `ClientProposeTxRequest` object, which
contains a `Tx`, and the client's minimum fee map digest. We introduce
`client_propose_tx_v2` which encrypts this `Request` object instead
of a naked `Tx`. This will come in handy if we want to add anything
else like this.

Clients will not be able to use `client_propose_tx_v2` until after the
V3 service is live, because existing consensus nodes don't have this
endpoint. However, in the next release of clients after the v3 release
is live, they could all start using `client_propose_tx_v2` and we could
deprecate / remove `client_propose_tx`.

I made the fog-sample-paykit exercise the `client_propose_tx_v2` route,
since the test client uses that and so we can see that it is working in
CD without making `mobilecoind` or anything else use it yet.