Make mc_ledger_db::Error implement Debug by forwarding to Display (#2673)
* Make mc_ledger_db::Error implement Debug by forwarding to Display

This may be controversial, but I think this may be a good change
to make in this particular case.

The goal is to fix #2666. A lot of our code takes a path to a
`LedgerDb` as a command line argument. Then it calls `LedgerDb::open`
on that path and unwraps any errors.

LMDB is a C library, and if it encounters any libc errors these get
returned as `lmdb::Error::Other(c_int)`.

If we Display this error, it will call `mdb_strerror`:

http://www.lmdb.tech/doc/group__mdb.html#ga569e66c1e3edc1a6016b86719ee3d098

Which will decide if it is an lmdb error or a libc error, and in the
latter case call `strerror` and in the former case do something
appropriate for the lmdb error.

If we print this error as `Debug`, it will do none of that, it just
gives us the `c_int`, and many developers aren't sure what to make
of `Lmdb(Other(2))` or `Lmdb(Other(13))`.

Sometimes they will start trying to match these numbers to this list:

https://docs.rs/lmdb-rkv/latest/lmdb/enum.Error.html

which is reasonable but will lead them totally down the wrong path.

The problem is that `unwrap` and `expect` both use the `Debug` output
for the error instead of the `Display` output.

We could try to rewrite all the unwraps of `LedgerDb` errors, but
(1) there a lot of them
(2) since the derived `Debug` output is so much worse than
the `Display` output, it seems like it might be better not derive
`Debug` and make `Debug` forward to `Display`.

It's hard for me to see a downside to that except that it's somewhat
unusual, and I see significant upside in this particular case.

* cargo fmt

* make sure the badrslot error is intelligible after this change