Tx std reshuffle (#2706)
* Rename transaction-std to transaction-builder

The name transaction-std was bad, because that crate really has
nothing to do with std. It doesn't even use std really.

Naming it `mc-transaction-builder` better scopes it to the purpose
of supporting the `TransactionBuilder` and related objects.

* make the transaction-builder crate no_std

This may be useful some day in the future, and it only requires
changing where we import symbols from. It's better for things to
be more easily portable than less, and not declare unncessary
dependencies, even on std.

* Update transaction/builder/README.md

Co-authored-by: Remoun Metyas <remoun@mobilecoin.com>

* rejigger no_std declaration per review

Co-authored-by: Remoun Metyas <remoun@mobilecoin.com>