Introduce an mc-transaction-extra crate taking bits of core and builder
The thing that got me thinking about this is:
* When trying to build the reserve auditor, the build fails unless
  we set SGX_MODE=HW and IAS_MODE, even though the reserve auditor
  does not do any attestation or check any IAS reports.
  I investigated and figured out that `mc-api` depends on
  `mc-transaction-builder` to get the `UnsignedTx` object,
  because that has a counterpart in `external.proto.
  I reported in #2630 that we should rejigger the dependency tree.
  (Actually, it's a separate question why transaction builder
  brings in the attest verifier code.)
* Initially my thought was that `UnsignedTx` should live in
  `mc-transaction-core` rather than `mc-transaction-builder`.
  However, that's actually not a great fix, because it forces the
  consensus-enclave-impl to depend on `UnsignedTx`, even though
  it doesn't use it.
* Still it seems reasonable that a project like reserve-auditor
  should be able to depend on `mc-api` without pulling in the
  `TransactionBuilder`, since it doesn't need to build any transactions.
  That feels like we are tightly coupling a huge amount of business
  logic to an API crate.

The more I thought about it, I became convinced that there needs to
be a third crate in here, because there are a bunch of types that,
are "important" and "relevant to transactions", and may even be
specified in MCIPs and involve cryptography etc. But, they aren't
directly needed for transaction validation, and they aren't part of
defining what is a valid mobilecoin transaction.

My opinion is that `mc-transaction-core` should be limited to just
the things that are required to be in the enclave for validation,
to the extent possible.

* This lets us put as little code in the enclave as possible (good)
* This lets us easily see that changing a particular struct cannot
  break enclave compatibility (good)
* During PR review, it becomes more clear that e.g. `InputRules`
  is part of the core and represents how "what is a valid transaction?"
  is changing, while `SignedContingentInput` is more like an "extra"
  schema, that isn't part of the core protocol, but rather some
  kind of layer 2 thing on top of the main protocol.

I identified a number of things that can move from `mc-transaction-core`
to this new `mc-transaction-extra` crate:

* TxOutConfirmationNumber
* SignedContingentInput

I also identified some things in `mc-transaction-builder` that could be
better off in the `mc-transaction-extra` crate:

* Memo types
* UnsignedTx

The memo builders remain in the builder crate.

This also fixes my original issue -- the `mc-api` crate no longer
depends on `mc-transaction-builder`.

If we merge this, I would rebase MCIP 52 work, and make it
so that:

* `TxSummary` (and friends) are part of `mc-transaction-core`
* `TxSummaryUnblindingData` (and friends) are part of `mc-transaction-extra`