Add `--check` option to rust lint.sh script (#2695)
* Add `--check` option to rust lint.sh script

Previously the lint.sh script would always check. Now the script will
attempt to fix and will only check if the `--check` flag is provided.

* Normalize the indent in `tools/lint.sh`

The previous update to `tools/lint.sh` brought in tab characters into
the file.  This resulted in mixed and space indent as well as
inconsistent indentation throughout the file.