add scheduled times and timezones for dependabot PRs
Currently, the dev-cluster can only build and run tests for a handful
of PRs at a time. If dependabot makes 7 PRs at once, this will swamp
the cluster, and all engineers will be blocked from iterating on CI
for several hours (even if CD is not run).

This change makes dependabot only create PRs for the root workspace
at midnight in the pacific timezone.

The other workspaces are assigned to 1 AM in the pacific timezone.

Given that most engineers are in one of the four US continental
timezones, this will always be between midnight and 5 AM for most
employees, so people won't have to kill the dependabot CI runs
manually in order to be productive.