Add TxSummary object and incorporate it into MLSAG signing digest (#2683)
* Add TxSummary object and incorporate it into MLSAG signing digest

The motivation here is that, hardware wallets like Ledger have
an expectation that the device can verify what is the balance
change that will result from signing a given transaction and who
the outbound money will be paid to.

However, currently we have scoped things so that hardware wallets
just are responsible for signing the Ring MLSAG, and they don't
see large parts of the Tx, including the TxPrefix.

The only way that the set of outputs is connected to the RingMLSAG
is that the TxPrefix is hashed into the "message", which is hashed
with more things to produce the "extended_message_digest", which
is the 32 byte digest that the RingMLSAG actually signs.

For the hardware wallet to really know what it is signing then,
it would have to reproduce the entire extended message digest.
The problem with this is that the TxPrefix can be > 100kb, which is
much larger than the few kb that the hardware device currently has
to handle, and is much too big to fit in the device's memory.

Instead, we introduce a `TxSummary` object and add an extra step
where it is hashed in just at the very end. The computer can then
prove that the hash the MLSAG is signing is connected to a given
TxSummary, and provide enough information from there to understand
the balance delta and where the money is going. This manages to
avoid sending all the merkle proofs for all the mixins and so
it can be done sending only a few kb.

The hope is to do this for v3, otherwise this issue may become
the long poll for shipping hardware wallet support.

* store input_rules with TxSummary, bring TxSummary in line with spec

* fixup

* Update transaction/core/src/tx_summary.rs

Co-authored-by: Remoun Metyas <remoun@mobilecoin.com>

* Update transaction/types/src/block_version.rs

Co-authored-by: Remoun Metyas <remoun@mobilecoin.com>

* Update transaction/core/src/tx_summary.rs

Co-authored-by: Remoun Metyas <remoun@mobilecoin.com>

* Update transaction/core/src/tx_summary.rs

Co-authored-by: Remoun Metyas <remoun@mobilecoin.com>

* Update transaction/core/src/tx_summary.rs

Co-authored-by: Remoun Metyas <remoun@mobilecoin.com>

* Update transaction/core/src/tx_summary.rs

Co-authored-by: Remoun Metyas <remoun@mobilecoin.com>

* fixups to idea for SCI support in TxSummary

* Update transaction/core/src/tx_summary.rs

Co-authored-by: Nick Santana <nick@mobilecoin.com>

* lint and clippy

* fixup TxOutSummary tags

* review comments, add types for digests, move digests to new module

* fixups to "associated_to_input_rules" test for TxOutSummary

* improved code comments and cargo fmt

Co-authored-by: Remoun Metyas <remoun@mobilecoin.com>
Co-authored-by: Nick Santana <nick@mobilecoin.com>