allow specifying a fog hint when minting (#2584)
* allow specifying a fog hint when minting

Consensus will now accept a MintTx that includes an encrypted fog
hint. This allows that e.g. a cross-chain bridge can be used directly
from a phone, and the cross-chain bridge is responsible for figuring
out the fog part (and not the mobilecoin consensus network).

* add support to mint client for minting to fog addresses

* add a block version check for the fog hint minting

* extend fog local network tests to try minting to fog accounts

This has the added benefit that we can exercise the atomic swap
stuff now in CI

* add a unit test

* cargo lock

* add more logging

* reduce number of tests created by test client

* fixup to ring sampling when matching SCIs

this prevents collisions across the sampled mixins for our own
inputs vs. the sci inputs

* try to fix more issues in ring sampling

* fixup

* fix cargo lock after rebase

* fix cargo lock again

* fix nick review comment

* Update consensus/mint-client/src/fog.rs

Co-authored-by: Nick Santana <nick@mobilecoin.com>

* Update consensus/mint-client/src/config.rs

Co-authored-by: Nick Santana <nick@mobilecoin.com>

* make e_fog_hint check part of mint tx validation, and tests

this resolves a Nick comment

* avoid duplicating grpcio environment code

* fix test build

* rename FogBits -> FogContext per discussion

Co-authored-by: Nick Santana <nick@mobilecoin.com>