Support parallel tx splitting in faucet worker, for better slamming (#2934)
* Support parallel tx splitting in faucet worker, for better slamming

Currently, the faucet worker submits at most one "splitting" Tx
at a time, whenever the queue of target-value Txs is not saturated.

This splitting Tx can produce 15 target value TxOut, and one change
TxOut.

Usually, the faucet has one massive TxOut which it got by bootstrap
or minting, and the worker is always splitting small TxOut's off
of this one.

This is fine when slam only has e.g. 500 TxOut's in it. Since the
worker thread can produce 15 txo/s sending Tx's one at a time, we
only have to wait about a minute to rebuild before slamming again.

However, James wants to do slams with 100,000 TxOut's like we did
when we did bootstrap-based slam, at least to see if there are
regressions.

Producing 100,000 TxOut's this way will take a very long time.
We have to submit 66,000 Tx's, and at one per second, that's
111 minutes. I have never had the patience to wait and test if
this works.

This change makes the faucet worker smarter. If it starts with
one massive TxOut, it first divides it into 16 about equal TxOuts.
Then whenever it needs more target value TxOuts, it sends 16 splitting
Tx's in parallel, each producing about 15 target value TxOut's,
continuously. With this improvement, it should take about 7 minutes
to be ready to do another slam.

I also added some logic to allow the faucet worker to defragment
itself a bit. This is because, in the above logic, it only looks
at the 16 largest UTXOs in its possession, and if these get too
small, it stops being able to do anything. However, frequently
slam is triggered where the faucet sends to itself, producing a
bunch of (target_value - fee) TxOut's. Eventually, it would get
stuck and not be able to slam anymore, which might be confusing
because it's balance is still pretty high.

The defragmentation check makes it able to still do something
with the lower value TxOut's if it's balance is totally fragmented.
This part doesn't work in parallel, but that should be okay for now.

* fixup comments

* fixups to defragmentation routine logic

* improve code comments

* Update mobilecoind-dev-faucet/src/worker.rs

Co-authored-by: Eran Rundstein <eran@rundste.in>

* Update mobilecoind-dev-faucet/src/worker.rs

Co-authored-by: Eran Rundstein <eran@rundste.in>

* Update mobilecoind-dev-faucet/src/worker.rs

Co-authored-by: Eran Rundstein <eran@rundste.in>

* Update mobilecoind-dev-faucet/src/worker.rs

Co-authored-by: Eran Rundstein <eran@rundste.in>

* address eran comment

* fix some nits, check on the status of in-flight Tx's unconditionally

* fix another remoun comment

* fix clippies

* make default queue size to be 10,000 and default slam size to be 10,000

* Make CI continue to do slams of size 500 only

* add a changelog entry

Co-authored-by: Eran Rundstein <eran@rundste.in>