Make mobilecoind, dev-faucet, fog-test-client use MRENCLAVE and support many css
This makes all the targets that create attestation verifiers in this repo
use MRENCLAVE for verification, and accept multiple CSS files (rather than
just one) as a CLI arg or env variable so that they can easily be made to
trust several versions of the enclave.

For things like dev-faucet and fog-test-client, this may simplify things
slightly operationally, since you won't have to rebuild against an old releases'
CSS files, you can simply install them with the deployed container.

For mobilecoind, this improves the trust model, since we're really using
SGX now to ensure that the correct code is running in the network.
It also ensures we're checking fog ingest using the MRENCLAVE validation.