WIP: make a way to store hardening advisory data with sigstructs and load it
This is meant to address comments on PR #3078, and allow that a client
can load multiple MRENCLAVE measurements on startup and also apply different
sets of hardening advisories to these as appropriate.

The idea is that there is a directory containing subdirectories with
`.css` sigstruct files and associated `.json` files. A client binary
can walk this directory and configure `MRENCLAVE` or `MRSIGNER` verifiers
using additional data obtained from the json files. We can ship the whole
bundle with the client, and it gets installed on the user's filesystem
with the app.

There is a lot of ambiguity here. Mostly we are optimizing towards,
reducing the amount of new tools we will have to create and use,
and particularly, avoiding creating a file format that embeds Intel's
sigstruct binary data.

If there is approval we will try to redevelop PR #3078 based on this.