Use "unknown" for `GIT_COMMIT` failures.
Previously `?????` would be returned for an inability to run git and
the git error would be returned for a failure running git, when trying
to determine the git commit during build. Now "unknown" will always be
returned in either failure condition.

The drive for this is to prevent log messages of the form
`"GIT_COMMIT" : "fatal: detected dubious "` which can happen in some CI
systems due to newer git `safe.directory` settings.