This comments out the "grafana push" part of the prometheus yml template.
AFAIK this grafana push hasn't been used by anyone in a few years,
and it requires you to do additional setup in grafana. If you just
want to talk to prometheus web interface, you have to comment this
out or you get logspam.

I feel like it makes sense to comment this out, but lmk if you'd rather
leave it in, not a big deal.