adjust readiness criteria for fog view
Jason reported that fog view is now declaring itself ready well
before the highest-processed-block-count is close to the block height.

In the past, the main bottleneck was loading the data from Postgres,
but now it's actually putting the data into ORAM.

In this commit, we refine the readiness criteria, so that both
the db fetcher and the enclave loading thread have to run out of
work before the server declares itself ready.