fog-view: Limit number of user events, and tell client this happened
Jason reported recently that fog test client failed in test-net with
a grpc maximum response size error. Eventually it recovered and I'm
not sure we fully determined the root cause.

One defect that we noticed during investigation is that, while the
client limits the number of ETxOutRecord's it requests, nothing limits
the number of user events that fog view returns -- if the client
starts from scratch, it always returns all the events.

This patch is meant to fix that. The SQL query is changed so that
there is a limit to how many user events it returns, and it is ordered
ascending by the user event id. This limit is configurable as a server
startup parameter. When the limit is reached, we flag to the client
that there may be more user events.

Then we update the `fog-view-protocol` polling module so that when
this flag is set, it keeps asking for more seeds until this flag
isn't set any more.

I believe that this is a backwards compatible change, because old
clients will ignore this flag, and for new clients talking to old
servers, this flag will have a default value of false, which matches
the status quo behavior of always returning all the results.