make a way to store hardening advisory data with enclave measurements and load it (#3148)
* make a way to store hardening advisory data with enclave measurements and load it

This is meant to address comments on PR #3078, and allow that a client
can load multiple MRENCLAVE measurements on startup and also apply different
sets of hardening advisories to these as appropriate.

The idea is that there is a file named e.g. `trusted-measurements.json`
which contains a json schema organizing enclave measurements, their respective
lists of mitigated hardening advisories, and what release they are associated
to, as well as any additional metadata for setting up MRSIGNER verifiers,
if that is desired.

Clients like mobilecoind or full-service can load this on startup to
configure their verifiers. SDKs could likely bake this in as a string literal,
or take it as a parameter.

This is more correct than what we are currently doing with passing CSS
files around, and it makes it easy to trust "previous release and next release".

If there is approval we will try to redevelop PR #3078 based on this.

* clippies

* adjust readme

* Update attest/verifier/config/README.md

Co-authored-by: Nick Santana <nick@mobilecoin.com>

* Update attest/verifier/config/README.md

Co-authored-by: Nick Santana <nick@mobilecoin.com>

* Update attest/verifier/config/README.md

Co-authored-by: Nick Santana <nick@mobilecoin.com>

* Update attest/verifier/config/src/lib.rs

Co-authored-by: Nick Santana <nick@mobilecoin.com>

* Update attest/verifier/config/src/lib.rs

Co-authored-by: Nick Santana <nick@mobilecoin.com>

* Update attest/verifier/config/src/lib.rs

Co-authored-by: Nick Santana <nick@mobilecoin.com>

* Update attest/verifier/config/src/lib.rs

Co-authored-by: Nick Santana <nick@mobilecoin.com>

* Update attest/verifier/config/src/lib.rs

Co-authored-by: Nick Santana <nick@mobilecoin.com>

* Update attest/verifier/config/Cargo.toml

Co-authored-by: Nick Santana <nick@mobilecoin.com>

* Update attest/verifier/config/src/lib.rs

Co-authored-by: Nick Santana <nick@mobilecoin.com>

* fix many review comments

also switch from serde_hex to the hex crate. the hex crate now
has a serde module, and seems like it's more actively maintained.
it also supports no_std.

also add some more tests of expected failures when parsing

* improve test

* fix cargo.lock

* fix cargo lock again

* Update attest/verifier/config/README.md

Co-authored-by: James Cape <james@mobilecoin.com>

* Update attest/verifier/config/Cargo.toml

Co-authored-by: James Cape <james@mobilecoin.com>

---------

Co-authored-by: Nick Santana <nick@mobilecoin.com>
Co-authored-by: James Cape <james@mobilecoin.com>