mobilecoind: fix a bug in swap generation
The `SignedContingentInputBuilder` does not know the TxOut global
indices of the ring elements, they usually aren't supplied to it
unless merkle proofs of membership were supplied to it.
It usually ends up setting `tx_out_global_indices` to a list of zeroes,
and it falls to the counterparty to find proofs of membership.

I forgot to do this in `generate_swap_impl`, so `mobilecoind` was
leaving `tx_out_global_indices` with an incorrect setting. This
commit fixes it and adds a test.