try to fix cargo downrevs in #3238
My method was:

(1) Check out HEAD~ Cargo.lock, and try to build.
(2) That didn't result in any fixes to the version numbers, which
    surprised me. So maybe something was yanked (???)
(3) Manually audit the diff from #3238 and make a list of everything
    that appeared to be down-reved.
(4) I constructed and then executed the following command:

```
cargo update -p digest@0.10.6 \
    -p semver@1.0.16 \
    -p cargo_metadata@0.15.3 \
    -p toml@0.5.11 \
    -p miniz_oxide@0.6.2 \
    -p object@0.30.3 \
    -p proc-macro2@1.0.51 \
    -p prost@0.11.8 \
    -p prost-derive@0.11.8 \
    -p regex@1.7.1 \
    -p reqwest@0.11.14 \
    -p cookie@0.16.2 \
    -p signal-hook@0.3.15 \
    -p wasm-bindgen@0.2.84 \
    -p link-cplusplus@1.0.8 \
    -p js-sys@0.3.61 \
    -p gimli@0.27.1
```

That produces this diff. I then tried to build to make sure this
looks right.