expose Schnorrkel Error from mc-crypto-keys
This type is already exposed in public APIs of RistrettoPublic and
various conversions, but it is not actually exported from the crate,
so there is no way for downstream to name the type.

This commit re-exports the type at crate level, which is already
part of the public API anyways.