fix enclave panic reporting so that it always shows the error
Most of the time when an enclave panics, particularly during local
development, the process dies with a hard abort before slog gets a
chance to flush it's buffer.

This commit makes it emit the panic message on STDERR first, which
means it will be in an OS buffer before the process ends.
We continue to write it as well as to slog, which writes to STDOUT.

I see no reason not to do this, it's a massive quality of life
improvement for anyone actually working on enclave-impl's.