refactor mobilecoind api to not depend on consensus-api
I am having a lot of build problems in another project due to
mc-mobilecoind-api pulling in mbedtls, which exports some symbols
that conflicts with another C library that some rust code depends on.

I determined that the only reason I have mbedtls in my build is
that `mobilecoind-api` depends on `consensus-api` and this pulls
in `mc-attest-core` etc.

This change allows me to break this dependency.