Add back config advisories
Add back the mitigated config advisories for the MRENCLAVE and
MRSIGNER trusted measurements. Previously the desire was to have a list
of advisories with a flag indicating what status was acceptable based on
the flags at
<https://api.portal.trustedservices.intel.com/documentation#pcs-tcb-info-model-v3>.

After some thought, changing here would require consumers to change from
their current use of 2 lists for each advisory type, while not providing
any benefit to said consumers.  The current logic of combining what's in
the config and hardening advisories for the
`ConfigurationAndSWHardeningNeeded` state will suffice.